+++ src/RcppExports.cpp
@@ -49,7 +49,7 @@ RcppExport SEXP _targeted_NB(SEXP ySEXP, SEXP xSEXP, SEXP xlevSEXP, SEXP ylevSEX
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -88,7 +88,7 @@ RcppExport SEXP _targeted_predNB(SEXP XSEXP, SEXP condprobSEXP, SEXP xordSEXP, S
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -125,7 +125,44 @@ RcppExport SEXP _targeted_ode_solve(SEXP fSEXP, SEXP inputSEXP, SEXP initSEXP, S
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
+    }
+    UNPROTECT(1);
+    return rcpp_result_gen;
+}
+// ode_solve2
+arma::mat ode_solve2(Rcpp::Function f, arma::mat input, arma::mat init, arma::mat par);
+static SEXP _targeted_ode_solve2_try(SEXP fSEXP, SEXP inputSEXP, SEXP initSEXP, SEXP parSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< Rcpp::Function >::type f(fSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type input(inputSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type init(initSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type par(parSEXP);
+    rcpp_result_gen = Rcpp::wrap(ode_solve2(f, input, init, par));
+    return rcpp_result_gen;
+END_RCPP_RETURN_ERROR
+}
+RcppExport SEXP _targeted_ode_solve2(SEXP fSEXP, SEXP inputSEXP, SEXP initSEXP, SEXP parSEXP) {
+    SEXP rcpp_result_gen;
+    {
+        Rcpp::RNGScope rcpp_rngScope_gen;
+        rcpp_result_gen = PROTECT(_targeted_ode_solve2_try(fSEXP, inputSEXP, initSEXP, parSEXP));
+    }
+    Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
+    if (rcpp_isInterrupt_gen) {
+        UNPROTECT(1);
+        Rf_onintr();
+    }
+    bool rcpp_isLongjump_gen = Rcpp::internal::isLongjumpSentinel(rcpp_result_gen);
+    if (rcpp_isLongjump_gen) {
+        Rcpp::internal::resumeJump(rcpp_result_gen);
+    }
+    Rboolean rcpp_isError_gen = Rf_inherits(rcpp_result_gen, "try-error");
+    if (rcpp_isError_gen) {
+        SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
+        UNPROTECT(1);
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -166,7 +203,7 @@ RcppExport SEXP _targeted_bin_logl(SEXP ySEXP, SEXP aSEXP, SEXP x1SEXP, SEXP x2S
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -207,7 +244,7 @@ RcppExport SEXP _targeted_bin_dlogl(SEXP ySEXP, SEXP aSEXP, SEXP x1SEXP, SEXP x2
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -246,7 +283,7 @@ RcppExport SEXP _targeted_bin_pa(SEXP ySEXP, SEXP aSEXP, SEXP x1SEXP, SEXP x2SEX
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -287,7 +324,7 @@ RcppExport SEXP _targeted_bin_dlogl_c(SEXP ySEXP, SEXP aSEXP, SEXP x1SEXP, SEXP 
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -329,7 +366,7 @@ RcppExport SEXP _targeted_bin_esteq(SEXP ySEXP, SEXP aSEXP, SEXP x1SEXP, SEXP x2
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -371,14 +408,14 @@ RcppExport SEXP _targeted_bin_esteq_c(SEXP ySEXP, SEXP aSEXP, SEXP x1SEXP, SEXP 
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
 }
 // ace_est
-Rcpp::List ace_est(const arma::vec& y, const arma::mat& a, const arma::mat& x1, const arma::mat& x2, const arma::vec& theta, const arma::vec& weights, bool binary);
-static SEXP _targeted_ace_est_try(SEXP ySEXP, SEXP aSEXP, SEXP x1SEXP, SEXP x2SEXP, SEXP thetaSEXP, SEXP weightsSEXP, SEXP binarySEXP) {
+Rcpp::List ace_est(const arma::vec& y, const arma::mat& a, const arma::mat& x1, const arma::mat& x2, const arma::vec& theta, const arma::vec& weights, const arma::vec& offset, std::string link);
+static SEXP _targeted_ace_est_try(SEXP ySEXP, SEXP aSEXP, SEXP x1SEXP, SEXP x2SEXP, SEXP thetaSEXP, SEXP weightsSEXP, SEXP offsetSEXP, SEXP linkSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::traits::input_parameter< const arma::vec& >::type y(ySEXP);
@@ -387,16 +424,17 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const arma::mat& >::type x2(x2SEXP);
     Rcpp::traits::input_parameter< const arma::vec& >::type theta(thetaSEXP);
     Rcpp::traits::input_parameter< const arma::vec& >::type weights(weightsSEXP);
-    Rcpp::traits::input_parameter< bool >::type binary(binarySEXP);
-    rcpp_result_gen = Rcpp::wrap(ace_est(y, a, x1, x2, theta, weights, binary));
+    Rcpp::traits::input_parameter< const arma::vec& >::type offset(offsetSEXP);
+    Rcpp::traits::input_parameter< std::string >::type link(linkSEXP);
+    rcpp_result_gen = Rcpp::wrap(ace_est(y, a, x1, x2, theta, weights, offset, link));
     return rcpp_result_gen;
 END_RCPP_RETURN_ERROR
 }
-RcppExport SEXP _targeted_ace_est(SEXP ySEXP, SEXP aSEXP, SEXP x1SEXP, SEXP x2SEXP, SEXP thetaSEXP, SEXP weightsSEXP, SEXP binarySEXP) {
+RcppExport SEXP _targeted_ace_est(SEXP ySEXP, SEXP aSEXP, SEXP x1SEXP, SEXP x2SEXP, SEXP thetaSEXP, SEXP weightsSEXP, SEXP offsetSEXP, SEXP linkSEXP) {
     SEXP rcpp_result_gen;
     {
         Rcpp::RNGScope rcpp_rngScope_gen;
-        rcpp_result_gen = PROTECT(_targeted_ace_est_try(ySEXP, aSEXP, x1SEXP, x2SEXP, thetaSEXP, weightsSEXP, binarySEXP));
+        rcpp_result_gen = PROTECT(_targeted_ace_est_try(ySEXP, aSEXP, x1SEXP, x2SEXP, thetaSEXP, weightsSEXP, offsetSEXP, linkSEXP));
     }
     Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
     if (rcpp_isInterrupt_gen) {
@@ -411,7 +449,7 @@ RcppExport SEXP _targeted_ace_est(SEXP ySEXP, SEXP aSEXP, SEXP x1SEXP, SEXP x2SE
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -449,7 +487,7 @@ RcppExport SEXP _targeted_fast_iid(SEXP ySEXP, SEXP pSEXP, SEXP x1SEXP, SEXP wei
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -483,7 +521,7 @@ RcppExport SEXP _targeted_clusterid(SEXP idSEXP) {
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -519,7 +557,7 @@ RcppExport SEXP _targeted_groupsum(SEXP xSEXP, SEXP clusterSEXP, SEXP reduceSEXP
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -555,7 +593,7 @@ RcppExport SEXP _targeted_softmax(SEXP lpSEXP, SEXP refSEXP, SEXP logSEXP) {
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -591,7 +629,41 @@ RcppExport SEXP _targeted_pava(SEXP ySEXP, SEXP xSEXP, SEXP weightsSEXP) {
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
+    }
+    UNPROTECT(1);
+    return rcpp_result_gen;
+}
+// nondom
+arma::mat nondom(const arma::mat& x);
+static SEXP _targeted_nondom_try(SEXP xSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< const arma::mat& >::type x(xSEXP);
+    rcpp_result_gen = Rcpp::wrap(nondom(x));
+    return rcpp_result_gen;
+END_RCPP_RETURN_ERROR
+}
+RcppExport SEXP _targeted_nondom(SEXP xSEXP) {
+    SEXP rcpp_result_gen;
+    {
+        Rcpp::RNGScope rcpp_rngScope_gen;
+        rcpp_result_gen = PROTECT(_targeted_nondom_try(xSEXP));
+    }
+    Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
+    if (rcpp_isInterrupt_gen) {
+        UNPROTECT(1);
+        Rf_onintr();
+    }
+    bool rcpp_isLongjump_gen = Rcpp::internal::isLongjumpSentinel(rcpp_result_gen);
+    if (rcpp_isLongjump_gen) {
+        Rcpp::internal::resumeJump(rcpp_result_gen);
+    }
+    Rboolean rcpp_isError_gen = Rf_inherits(rcpp_result_gen, "try-error");
+    if (rcpp_isError_gen) {
+        SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
+        UNPROTECT(1);
+        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -604,18 +676,20 @@ static int _targeted_RcppExport_validate(const char* sig) {
         signatures.insert("Rcpp::List(*.NB)(arma::vec,arma::mat,arma::uvec,arma::vec,arma::vec,double)");
         signatures.insert("arma::mat(*.predNB)(arma::mat const&,Rcpp::List const&,Rcpp::List const&,arma::uvec,arma::vec,double)");
         signatures.insert("arma::mat(*.ode_solve)(odeptr_t,arma::mat,arma::mat,arma::mat)");
+        signatures.insert("arma::mat(*.ode_solve2)(Rcpp::Function,arma::mat,arma::mat,arma::mat)");
         signatures.insert("arma::vec(*bin_logl)(const arma::vec&,const arma::vec&,const arma::mat&,const arma::mat&,const arma::vec,const arma::vec&,std::string,bool)");
         signatures.insert("arma::mat(*bin_dlogl)(const arma::vec&,const arma::vec&,const arma::mat&,const arma::mat&,const arma::vec,const arma::vec&,std::string,bool)");
         signatures.insert("arma::mat(*bin_pa)(const arma::vec&,const arma::vec&,const arma::mat&,const arma::mat&,const arma::vec,std::string)");
         signatures.insert("arma::cx_mat(*bin_dlogl_c)(const arma::cx_vec&,const arma::cx_vec&,const arma::cx_mat&,const arma::cx_mat&,const arma::cx_vec,const arma::cx_vec&,std::string,bool)");
         signatures.insert("arma::mat(*bin_esteq)(const arma::vec&,const arma::vec&,const arma::mat&,const arma::mat&,const arma::vec&,arma::vec,arma::vec,const arma::vec&,std::string)");
         signatures.insert("arma::cx_mat(*bin_esteq_c)(const arma::cx_vec&,const arma::cx_vec&,const arma::cx_mat&,const arma::cx_mat&,const arma::cx_mat&,arma::cx_vec,arma::cx_vec,const arma::cx_vec&,std::string)");
-        signatures.insert("Rcpp::List(*ace_est)(const arma::vec&,const arma::mat&,const arma::mat&,const arma::mat&,const arma::vec&,const arma::vec&,bool)");
+        signatures.insert("Rcpp::List(*ace_est)(const arma::vec&,const arma::mat&,const arma::mat&,const arma::mat&,const arma::vec&,const arma::vec&,const arma::vec&,std::string)");
         signatures.insert("arma::mat(*fast_iid)(const arma::vec&,const arma::vec&,const arma::mat&,const arma::vec&,bool)");
         signatures.insert("List(*.clusterid)(const arma::uvec&)");
         signatures.insert("arma::mat(*.groupsum)(const arma::mat&,const arma::uvec&,bool)");
         signatures.insert("arma::mat(*.softmax)(arma::mat&,bool,bool)");
         signatures.insert("List(*.pava)(const arma::vec&,const NumericVector&,const NumericVector&)");
+        signatures.insert("arma::mat(*.nondom)(const arma::mat&)");
     }
     return signatures.find(sig) != signatures.end();
 }
@@ -625,6 +699,7 @@ RcppExport SEXP _targeted_RcppExport_registerCCallable() {
     R_RegisterCCallable("targeted", "_targeted_.NB", (DL_FUNC)_targeted_NB_try);
     R_RegisterCCallable("targeted", "_targeted_.predNB", (DL_FUNC)_targeted_predNB_try);
     R_RegisterCCallable("targeted", "_targeted_.ode_solve", (DL_FUNC)_targeted_ode_solve_try);
+    R_RegisterCCallable("targeted", "_targeted_.ode_solve2", (DL_FUNC)_targeted_ode_solve2_try);
     R_RegisterCCallable("targeted", "_targeted_bin_logl", (DL_FUNC)_targeted_bin_logl_try);
     R_RegisterCCallable("targeted", "_targeted_bin_dlogl", (DL_FUNC)_targeted_bin_dlogl_try);
     R_RegisterCCallable("targeted", "_targeted_bin_pa", (DL_FUNC)_targeted_bin_pa_try);
@@ -637,6 +712,7 @@ RcppExport SEXP _targeted_RcppExport_registerCCallable() {
     R_RegisterCCallable("targeted", "_targeted_.groupsum", (DL_FUNC)_targeted_groupsum_try);
     R_RegisterCCallable("targeted", "_targeted_.softmax", (DL_FUNC)_targeted_softmax_try);
     R_RegisterCCallable("targeted", "_targeted_.pava", (DL_FUNC)_targeted_pava_try);
+    R_RegisterCCallable("targeted", "_targeted_.nondom", (DL_FUNC)_targeted_nondom_try);
     R_RegisterCCallable("targeted", "_targeted_RcppExport_validate", (DL_FUNC)_targeted_RcppExport_validate);
     return R_NilValue;
 }
@@ -647,18 +723,20 @@ static const R_CallMethodDef CallEntries[] = {
     {"_targeted_NB", (DL_FUNC) &_targeted_NB, 6},
     {"_targeted_predNB", (DL_FUNC) &_targeted_predNB, 6},
     {"_targeted_ode_solve", (DL_FUNC) &_targeted_ode_solve, 4},
+    {"_targeted_ode_solve2", (DL_FUNC) &_targeted_ode_solve2, 4},
     {"_targeted_bin_logl", (DL_FUNC) &_targeted_bin_logl, 8},
     {"_targeted_bin_dlogl", (DL_FUNC) &_targeted_bin_dlogl, 8},
     {"_targeted_bin_pa", (DL_FUNC) &_targeted_bin_pa, 6},
     {"_targeted_bin_dlogl_c", (DL_FUNC) &_targeted_bin_dlogl_c, 8},
     {"_targeted_bin_esteq", (DL_FUNC) &_targeted_bin_esteq, 9},
     {"_targeted_bin_esteq_c", (DL_FUNC) &_targeted_bin_esteq_c, 9},
-    {"_targeted_ace_est", (DL_FUNC) &_targeted_ace_est, 7},
+    {"_targeted_ace_est", (DL_FUNC) &_targeted_ace_est, 8},
     {"_targeted_fast_iid", (DL_FUNC) &_targeted_fast_iid, 5},
     {"_targeted_clusterid", (DL_FUNC) &_targeted_clusterid, 1},
     {"_targeted_groupsum", (DL_FUNC) &_targeted_groupsum, 3},
     {"_targeted_softmax", (DL_FUNC) &_targeted_softmax, 3},
     {"_targeted_pava", (DL_FUNC) &_targeted_pava, 3},
+    {"_targeted_nondom", (DL_FUNC) &_targeted_nondom, 1},
     {"_rcpp_module_boot_riskregmodel", (DL_FUNC) &_rcpp_module_boot_riskregmodel, 0},
     {"_targeted_RcppExport_registerCCallable", (DL_FUNC) &_targeted_RcppExport_registerCCallable, 0},
     {NULL, NULL, 0}